readme and facilitate setup
